[AAC/import] AAC/ADTS importer, still some seeking issue

+++ avidemux_core/ADM_coreAudio/src/ADM_audioClock.cpp
@@ -38,7 +38,7 @@ bool           audioClock::advanceBySample(uint32_t samples)
  * */
 uint64_t       audioClock::getTimeUs(void)
 {
-float f=_nbSamples;
+double  f=_nbSamples;
         f=f*1000*1000;
         f/=_frequency;
         return _baseClock+(uint64_t)(f+0.5);
@@ -49,7 +49,7 @@ float f=_nbSamples;
 bool           audioClock::setTimeUs(uint64_t clk)
 {
         uint64_t curTime=getTimeUs();
-        int64_t delta=(int64_t)clk-(int64_t)curTime;
+        int64_t  delta=(int64_t)clk-(int64_t)curTime;
         if(labs((long int)delta)<2000) return true;
         printf("[audioClock] Drift detected :%"PRIu64" vs %"PRIu64", delta=%"PRId64"\n",curTime,clk,delta);
         _nbSamples=0;
